functions/predictor/requirements.txt
functions-framework
google-cloud-storage
google-cloud-firestore
pandas
numpy
scikit-learn
xgboost